Move the flag registration code from pkg/apis/componentconfig to pkg/util/flag
